Nail: Applied the previous solution to feedback I and III to all three letters.
The problem with the number of differences between the correct and user answer
when only one and correct item is listed is solved concurrently with item
IV of the feedback which calls for all items to be entered at once
before generating hint or feedback to the user. For the first letter
only at this point.